// client/packages/big-advancedsearch/src/common/searchresult.ts
/*********************************************************************************
 * Copyright (c) 2023 borkdominik and others.
 *
 * This program and the accompanying materials are made available under the
 * terms of the MIT License which is available at https://opensource.org/licenses/MIT.
 *
 * SPDX-License-Identifier: MIT
 *********************************************************************************/

export interface SearchResult {
    id: string;
    type: string;
    name: string;
    parentName?: string;
    details?: string;
}